Added working C library dll

+++ calculator.c
@@ -1,22 +1,22 @@
 #include "calculator.h"
 #include <stdio.h>
 
-EXPORT double add(double num_a, double num_b)
+EXPORT float add(float num_a, float num_b)
 {
     return num_a + num_b;
 }
 
-EXPORT double sub(double num_a, double num_b)
+EXPORT float sub(float num_a, float num_b)
 {
     return num_a + num_b;
 }
 
-EXPORT double mul(double num_a, double num_b)
+EXPORT float mul(float num_a, float num_b)
 {
     return num_a * num_b;
 }
 
-EXPORT double div(double num_a, double num_b)
+EXPORT float div(float num_a, float num_b)
 {
     if(num_b == 0)
     {
@@ -29,12 +29,12 @@ EXPORT double div(double num_a, double num_b)
     }
 }
 
-EXPORT int mod(double num_a, double num_b)
+EXPORT float mod(float num_a, float num_b)
 {
-    return (int)num_a % (int)num_b;
+    return (float)((int)num_a % (int)num_b);
 }
 
-EXPORT double calculator(char operator, double num_a, double num_b)
+EXPORT float calculator(char operator, float num_a, float num_b)
 {
     switch (operator)
     {
